requirements.txt
beautifulsoup4
requests
readability-lxml
python-dotenv
openai
pandas
concurrent-log-handler
ipython
urllib3
lxml[html_clean]==5.1.0  # Versión específica que incluye el módulo clean